% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_activity.R
\name{get_account_activity}
\alias{get_account_activity}
\title{Get the activity in an account from a given start and end date, including buys, sells, and other money movements like dividends.}
\usage{
get_account_activity(account_number, start_date, end_date)
}
\arguments{
\item{account_number}{a user's M1 Finance account number.}

\item{start_date}{a string containing the starting date for the activity export, must not be greater than 2 years before the end date.}

\item{end_date}{a string containing the end date for the activity export.}
}
\description{
Get the activity in an account from a given start and end date, including buys, sells, and other money movements like dividends.
}
