% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{get_APEX_auth_token}
\alias{get_APEX_auth_token}
\title{Get an authorized token to use during your R session.}
\usage{
get_APEX_auth_token(username, password)
}
\arguments{
\item{username}{the username for your APEX Clearing account.}

\item{password}{the password for your APEX Clearing account.}
}
\value{
a global environment variable called APEX_token which is used in the other helper functions.
}
\description{
A token is required in order to make requests, this token is hidden in the cookies generated when a user logs into their account.
This function uses the login service and grabs the token from the cookies returned when the request is successful.
}
\details{
If a user is uncomfortable writing out their username/password in an R file as plaintext, the function can be left blank and
RStudio will prompt the user to enter their username and password. This is not stored anywhere once the function is finished running - only the AUTH_token is.
}
\examples{
\dontrun{
get_APEX_auth_token(username = "username", password = "password")
get_APEX_auth_token() # RStudio prompts the user for their login
}
}
