% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{get_credentials}
\alias{get_credentials}
\title{Get credentials to obtain an APEX auth token.}
\usage{
get_credentials()
}
\value{
a named list with a user's credentials
}
\description{
Get credentials to obtain an APEX auth token.
}
\details{
This function checks whether a session is interactive. If it is, it prompts the
user to enter their username and password. Otherwise, it checks .Renviron for the
credentials.
}
