% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gain_loss_by_term}
\alias{get_gain_loss_by_term}
\title{Helper function to clean up code which generates realized gains and losses.}
\usage{
get_gain_loss_by_term(json_response, short_or_long = c("short", "long"))
}
\arguments{
\item{json_response}{the API response in JSON format.}

\item{short_or_long}{one of either "short" or "long" for the term in which gains are evaluated.}
}
\description{
Helper function to clean up code which generates realized gains and losses.
}
