% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gains-and-losses.R
\name{get_open_positions}
\alias{get_open_positions}
\title{Get the open positions in your portfolio.}
\usage{
get_open_positions(account_number)
}
\arguments{
\item{account_number}{a user's M1 Finance account number.}
}
\description{
Get the open positions in your portfolio.
}
