% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_positions.R
\name{get_portfolio_positions}
\alias{get_portfolio_positions}
\title{Get the current positions in your portfolio.}
\usage{
get_portfolio_positions(account_number)
}
\arguments{
\item{account_number}{a user's M1 Finance account number.}
}
\description{
Get the current positions in your portfolio.
}
\details{
This function returns top-level information about a user's portfolio, including the market value
and the number of shares they own.
}
