% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gains-and-losses.R
\name{get_realized_gains_losses}
\alias{get_realized_gains_losses}
\title{Get realized gains and losses for a given portfolio.}
\usage{
get_realized_gains_losses(account_number)
}
\arguments{
\item{account_number}{a user's M1 Finance account number.}
}
\description{
Get realized gains and losses for a given portfolio.
}
